cads/requirements.txt
# CADS Data Pipeline Dependencies

# Core Dependencies
pandas>=1.5.0
numpy>=1.21.0
python-dotenv>=0.19.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0

# Machine Learning
scikit-learn>=1.1.0
sentence-transformers>=2.2.0
umap-learn>=0.5.3
hdbscan>=0.8.28

# API Clients
requests>=2.28.0
groq>=0.4.0

# Utilities
tqdm>=4.64.0
joblib>=1.2.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Development & Testing
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0

# Optional: Jupyter for analysis
jupyter>=1.0.0
ipykernel>=6.15.0